requirements.txt
yfinance
pandas
pytz
requests
pyyaml